refactor(#188): 유저 쿠폰 라우터 이동 및 리팩토링
