+++ src/app/mypage/user/[coupon-pass]/_components/ClassFilterSelectUser.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { FilterSVG, ResetSVG } from '@/icons/svg';
 
 interface ClassFilterSelectUserProps {
